:sparkles: Write Controller::update_discrete_states
It requires a virtual pure method 'compute_command' that should be
specified in each specific controller (PIDController for example).

The continuous system argument of update_discrete_states should always be a 'Sim*' instance,
because the setpoint and measured states need to get their inputs from a Sim.
As 'ssc:::solver::DiscreteSystem' is more generic and does not know 'sim' yet,
the continuous system argument is typed as a 'ssc::solver::ContinuousSystem' but will
be casted into 'Sim*'.