:white_check_mark: Disabled unit test SimTest.LONG_bug_2732.
The changes to Sim::initialize_system_outputs_before_first_observation no longer allow a simulation to state with invalid states, even if there are no more time steps to compute.

However this should be investigated further as the simulation in this test only diverges on Windows, but not on Linux.