:bug: Fix CI problem steming from 'interfaces' repo
Needed to add '--fix-missing' to apt-get update in interfaces's Dockerfiles.
Cf.
https://gitlab.com/sirehna_naval_group/sirehna/xdyn/-/jobs/1631332959
for the build error in xdyn, which is similar to the build error
https://gitlab.com/sirehna_naval_group/sirehna/interfaces/-/jobs/1624892335
in 'interfaces'. Used
https://github.com/cookiecutter-flask/cookiecutter-flask/issues/736#issue-599215458
to solve it.
This problem only happened on Gitlab.com, not SIREHNA's internal Gitlab
server, but I suspect it might happen again, which is why I added
--fix-missing in all interfaces's Dockerfiles.