:bug: Fixed display bug detected in ELRPANG
Using abs meant the value was converted to an int.
We don't even need the absolute value here, but if
we did, using std::abs would also solve the problem.
We took the opportunity to increase the number of
digits.

We were getting absurd warnings such as:

xdyn-cs_1                             | WARNING: The encounter period
Te=62s is outside of the range [4.18879,62.8319]s provided in the HDB
file. The response will be interpolated outside the bounds.