:sparkles: Added Observer::flush
Because the serialization will now take place in two steps (one before
the solver update & one after), the order the variables will be
serialized does not necessarily match the requested order in the outputs
section. Hence we cache all serialization until we have them all & then
we just flush the outputs to a file.
For HDF5 files, this might not be necessary as the path to each variable
written in the file is unique, hence the order should not matter.