:sparkles: Changed the way observations are done in xdyn
So far, we were observing forces after the solver step. This meant the
forces we were observing had been calculated at an intermediate timestep
(for Euler, it's even worse: they were outputted at t but the values
corresponded to t+dt).
Now we separate things that can be serialized before a solver step (such
as forces & discrete states) and things that need to be serialized after
the solver step (continuous states, mainly). We use a cache to make sure
we don't compute the same forces twice in a row.