:sparkles: Add gRPC Python interface modules to xdyn Docker image
This is so those files are in the wave-server-airy Docker image (built
in poseidon/wass), which, in turn, is used in elrpang.
It means we are sure of using a version of those files that is coherent
between xdyn and the other services.

Cf. https://gitlab.sirehna.com/wass/elrpang/-/blob/62-interface-entre-l-assimilation-et-la-queue-de-message-partie-2/test/Dockerfile#L1